chore: Change function interface
