+++ lib/baha_crawler.js
@@ -31,16 +31,18 @@ const self = {
     },
 
     /**
-     *  @param {number} num
-     *  @param {boolean} skipTPs
+     *  @param {string}     options.board
+     *  @param {number}     options.pages
+     *  @param {boolean}    options.skipTPs
      *  @returns {object}
      */
-    getResults: async (board = '23805', num = 1, skipTPs = true) => {
+    getResults: async (options) => {    
         let data_pages = [];
         let numOfPage = 1;
-        self.scrapingBoard = board;
-        self.scrapingPages = (num < 0) ? 1 : num;
-        self.skipTopPosts = skipTPs;
+        options = options || {};
+        self.scrapingBoard = options.board || '23805';
+        self.scrapingPages = (options.pages < 0) ? 1 : options.pages;
+        self.skipTopPosts = options.skipTPs && true;
         
         // 巴哈網頁新增縮圖功能後，如果要回到原來的列表，要加入以下cookie
         const cookie = {
@@ -54,7 +56,7 @@ const self = {
         }
 
         /***** 前往 baha要爬的版面的 Url (最新頁面) *****/
-        const bahaUrl = 'https://forum.gamer.com.tw/B.php?page=1&bsn=' + board;
+        const bahaUrl = 'https://forum.gamer.com.tw/B.php?page=1&bsn=' + self.scrapingBoard;
         self.page.setViewport({
             width: 1024,
             height: 768
@@ -68,12 +70,12 @@ const self = {
             self.browser.close();
         }
 
-        data_pages.push(await self.page.evaluate(self._scrapingOnePage, skipTPs));
+        data_pages.push(await self.page.evaluate(self._scrapingOnePage, self.skipTopPosts));
 
         for (let i = 1; i < self.scrapingPages; i++) {
             numOfPage++;
             /***** 前往下一頁 *****/
-            let pageUrl = 'https://forum.gamer.com.tw/B.php?page=' + numOfPage.toString() + '&bsn=' + board;
+            let pageUrl = 'https://forum.gamer.com.tw/B.php?page=' + numOfPage.toString() + '&bsn=' + self.scrapingBoard;
             try {
                 await self.page.goto(pageUrl);
                 await self.page.waitForSelector(self.stopSelector);
@@ -82,7 +84,7 @@ const self = {
                 await self.browser.close();
             }
             /***** 抓取網頁資料 (下一頁) *****/
-            data_pages.push(await self.page.evaluate(self._scrapingOnePage, skipTPs));
+            data_pages.push(await self.page.evaluate(self._scrapingOnePage, self.skipTopPosts));
         }
 
         /***** 將多頁資料 "照實際新舊順序" 合成 1 個物件 *****/
