+++ package.json
@@ -1,27 +1,35 @@
 {
   "name": "baha-crawler",
-  "version": "1.0.0",
-  "description": "baha-crawler.js is a web crawler module designed to scarp data from Bahamut forum.",
+  "version": "2.0.0",
+  "description": "baha-crawler is a web crawler module designed to scarp data from Bahamut forum.",
   "main": "index.js",
+  "files": [
+    "index.js",
+    "lib/**"
+  ],
   "directories": {
     "lib": "lib"
   },
   "scripts": {
+    "start": "node examples/demo.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/WayneChang65/baha-crawler.js.git"
+    "url": "git+https://github.com/WayneChang65/baha-crawler.git"
   },
   "keywords": [
-    "baha"
+    "baha",
+    "crawler",
+    "scrape",
+    "spider"
   ],
   "author": "WayneChang65",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/WayneChang65/baha-crawler.js/issues"
+    "url": "https://github.com/WayneChang65/baha-crawler/issues"
   },
-  "homepage": "https://github.com/WayneChang65/baha-crawler.js#readme",
+  "homepage": "https://github.com/WayneChang65/baha-crawler#readme",
   "dependencies": {
     "puppeteer": "^1.14.0"
   }
